Change Package to "org.eclipse.velocitas.sdk"
