+++ sdk/src/test/java/org/eclipse/velocitas/kotest/KotestProjectConfig.kt
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.eclipse.velocitas.kotest
+
+import io.kotest.core.config.AbstractProjectConfig
+
+// https://kotest.io/docs/framework/project-config.html
+object KotestProjectConfig : AbstractProjectConfig() {
+    override var displayFullTestPath: Boolean? = true
+}
